+++ packages/apollo-utilities/src/directives.ts
@@ -2,10 +2,12 @@
 // the `skip` and `include` directives within GraphQL.
 import {
   FieldNode,
+  OperationDefinitionNode,
   SelectionNode,
   VariableNode,
   BooleanValueNode,
   DirectiveNode,
+  DocumentNode,
 } from 'graphql';
 
 import { argumentsObjectFromField } from './storeUtils';
@@ -91,4 +93,57 @@ export function shouldInclude(
   });
 
   return res;
+}
+
+export function flattenSelections(selection: SelectionNode): SelectionNode[] {
+  if (
+    !(selection as FieldNode).selectionSet ||
+    !((selection as FieldNode).selectionSet.selections.length > 0)
+  )
+    return [selection];
+
+  return [selection].concat(
+    (selection as FieldNode).selectionSet.selections
+      .map(selectionNode =>
+        [selectionNode].concat(flattenSelections(selectionNode)),
+      )
+      .reduce((selections, selected) => selections.concat(selected), []),
+  );
+}
+
+const added = new Map();
+export function getDirectiveNames(doc: DocumentNode) {
+  const cached = added.get(doc);
+  if (cached) return cached;
+
+  // operation => [names of directives];
+  const directives = doc.definitions
+    .filter(
+      (definition: OperationDefinitionNode) =>
+        definition.selectionSet && definition.selectionSet.selections,
+    )
+    // operation => [[Selection]]
+    .map(x => flattenSelections(x as any))
+    // [[Selection]] => [Selection]
+    .reduce((selections, selected) => selections.concat(selected), [])
+    // [Selection] => [Selection with Directives]
+    .filter(
+      (selection: SelectionNode) =>
+        selection.directives && selection.directives.length > 0,
+    )
+    // [Selection with Directives] => [[Directives]]
+    .map((selection: SelectionNode) => selection.directives)
+    // [[Directives]] => [Directives]
+    .reduce((directives, directive) => directives.concat(directive), [])
+    // [Directives] => [Name]
+    .map((directive: DirectiveNode) => directive.name.value);
+
+  added.set(doc, directives);
+  return directives;
+}
+
+export function hasDirectives(names: string[], doc: DocumentNode) {
+  return getDirectiveNames(doc).some(
+    (name: string) => names.indexOf(name) > -1,
+  );
 }