+++ packages/apollo-utilities/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change log
 
 ### vNEXT
+- Added hasDirectives to recurse the AST and return a boolean for an array of directive names
 - Improved performance of common store actions by memoizing addTypename and removeConnectionDirective